# [dask] speed up tests