@trivialfis 2021-06-03
The slow test is probably not caused by dask but by hypothesis.