@jmoralez 2021-06-04
I see. The total runtime for the tests in this PR was [14 minutes](https://xgboost-ci.net/blue/organizations/jenkins/xgboost/detail/PR-7020/1/pipeline/197#step-225-log-2154) and the current master is around [17 minutes](https://xgboost-ci.net/blue/organizations/jenkins/xgboost/detail/master/878/pipeline/207#step-237-log-1542). Should I add the `client` fixture to tests that currently build a cluster internally like these?
https://github.com/dmlc/xgboost/blob/7beb2f7fae9589750c8d8dd01d039f90c71fc81f/tests/python/test_with_dask.py#L67
https://github.com/dmlc/xgboost/blob/7beb2f7fae9589750c8d8dd01d039f90c71fc81f/tests/python/test_with_dask.py#L110
That could probably reduce the runtime by a couple more minutes. Or if you have any other suggestions for maybe improving the hypothesis ones I could look into it.