@trivialfis 2021-06-04
> hould I add the client fixture to tests that currently build a cluster internally like these?

Some tests use specific number of workers so that they have to define their own cluster.

> Or if you have any other suggestions for maybe improving the hypothesis ones I could look into it.

Sorry I don't have any suggestion, you know these better than me. ;-)