@jmoralez 2021-06-04
I ran the tests that take the most time and pretty much all the time is spent training so I don't think there's something that could be improved there. There are a couple of clusters that get created with `kWorkers` and no `threads_per_worker`, I could create an additional fixture like `kCluster` and pass that to those tests so that they reuse that one.

This didn't improve as much as I hoped haha so please let me know if its useful at all.