@trivialfis 2021-06-05
> I ran the tests that take the most time and pretty much all the time is spent training so I don't think there's something that could be improved there.

That's fine, the hypothesis tests take longer than we would like but highly effective at catching bugs.

> There are a couple of clusters that get created with kWorkers and no threads_per_worker

Thank you for looking into them!

>  could create an additional fixture like kCluster and pass that to those tests so that they reuse that one.

I will follow up on making those changes since I wrote most of  the tests, I should cleanup my own mess.

> This didn't improve as much as I hoped haha so please let me know if its useful at all.

Of course it's useful and thank you!  After merging the PR we know that we should focus on other places.